feat(...) : startActivityForResult, optionItemSelected 공지사항 뷰로 이동
