+++ app/src/main/java/com/iron/espresso/presentation/home/mystudy/studydetail/NoticeFragment.kt
@@ -5,17 +5,21 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuInflater
+import android.view.MenuItem
 import android.view.View
+import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.fragment.app.viewModels
 import com.iron.espresso.R
 import com.iron.espresso.base.BaseFragment
+import com.iron.espresso.data.model.NoticeItem
 import com.iron.espresso.databinding.FragmentNoticeBinding
 import com.iron.espresso.model.response.notice.NoticeListResponse
 import com.iron.espresso.model.response.notice.NoticeResponse
 import com.iron.espresso.presentation.home.mystudy.StudyDetailActivity.Companion.DEFAULT_VALUE
 import com.iron.espresso.presentation.home.mystudy.StudyDetailActivity.Companion.STUDY_ID
 import com.iron.espresso.presentation.home.mystudy.adapter.NoticeAdapter
+import com.iron.espresso.presentation.home.mystudy.studydetail.NoticeCreateActivity.Companion.NOTICE_ITEM
 
 class NoticeFragment : BaseFragment<FragmentNoticeBinding>(R.layout.fragment_notice) {
 
@@ -51,7 +55,7 @@ class NoticeFragment : BaseFragment<FragmentNoticeBinding>(R.layout.fragment_not
                     noticeItem.id,
                     studyId
                 )
-            }, REQUEST_CODE)
+            }, REQUEST_DELETE_CODE)
         }
 
         binding.noticeList.adapter = noticeAdapter
@@ -60,10 +64,12 @@ class NoticeFragment : BaseFragment<FragmentNoticeBinding>(R.layout.fragment_not
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
 
-        if (requestCode == REQUEST_CODE && resultCode == RESULT_OK) {
+        if (requestCode == REQUEST_DELETE_CODE && resultCode == RESULT_OK) {
             viewModel.showNoticeList(studyId)
+        } else if (requestCode == REQUEST_CREATE_CODE && resultCode == RESULT_OK) {
+            val noticeItem = data?.getSerializableExtra(NOTICE_ITEM) as NoticeItem
+            Toast.makeText(context, "$noticeItem", Toast.LENGTH_SHORT).show()
         }
-
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
@@ -71,8 +77,28 @@ class NoticeFragment : BaseFragment<FragmentNoticeBinding>(R.layout.fragment_not
         super.onCreateOptionsMenu(menu, inflater)
     }
 
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            android.R.id.home -> {
+                activity?.finish()
+            }
+            R.id.create_notice -> {
+                startActivityForResult(
+                    NoticeCreateActivity.getInstance(requireContext()),
+                    REQUEST_CREATE_CODE
+                )
+            }
+
+            else -> {
+                Toast.makeText(context, "${item.title}", Toast.LENGTH_SHORT).show()
+            }
+        }
+        return true
+    }
+
     companion object {
-        private const val REQUEST_CODE = 1
+        private const val REQUEST_DELETE_CODE = 1
+        private const val REQUEST_CREATE_CODE = 2
         fun newInstance(data: Int) =
             NoticeFragment().apply {
                 arguments = Bundle().apply {
